Add script _review/run_screen_qac
This is a helper script to start screen with multiple windows using
`openqa-revtui` and the corresponding toml files for QAC.
The first windows will be a editor opend for notes. The name of the file
contains the year and the week number. The note file is places in the
same folder as the `run_screen_qac` script.

Usage:
```
ln -s ~/openqa-mon/_review/run_screen_qac ~/bin/openqa-revtui-qac
PATH=~/bin:$PATH
openqa-revtui-qac
```